backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Cors==3.0.10
Werkzeug==2.2.2
PyJWT==2.10.1